+++ eq-author/src/storybook/Components-themed/InputONS.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas } from "@storybook/addon-docs/blocks";
-import Input from "components-themed/Input";
+import Input from "components-themed/input";
 import { Label } from "components/Forms";
 import InlineField from "components/AnswerContent/Format/InlineField";
 import Theme from "contexts/themeContext";
